+++ plat/nxp/s32/s32g/s32g_mc_me.c
@@ -1,12 +1,12 @@
 /*
- * Copyright 2019-2021 NXP
+ * Copyright 2019-2022 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 #include <lib/mmio.h>
 #include "s32g_mc_me.h"
 
-void s32g_set_stby_master_core(uint8_t part, uint8_t core)
+void s32g_set_stby_master_core(uint8_t part, unsigned int core)
 {
 	/* Set the master core for the standby sequence */
 	mmio_write_32(MC_ME_MAIN_COREID,
